Fix PCRE for splitting email Reference headers
Fixes cases where email clients do not separate the *references* in the header with a simple space, or space character. The RFC somehow even allows textual comments between references. By simply splitting on everything not a closing angle bracket after an open angle bracket (non-greedy) this should work and split the string to an array, as expected, for all edge cases.